+++ docassemble_webapp/docassemble/webapp/backend.py
@@ -39,6 +39,7 @@ import docassemble.webapp.worker
 from docassemble.webapp.mailgun_mail import Mail as MailgunMail
 from docassemble.webapp.sendgrid_mail import Mail as SendgridMail
 from docassemble.webapp.fixpickle import fix_pickle_obj, fix_pickle_dict
+from docassemble.webapp.screenreader import to_text
 import pandas
 import math
 import xml.etree.ElementTree as ET
@@ -312,7 +313,8 @@ docassemble.base.functions.update_server(default_language=DEFAULT_LANGUAGE,
                                          server_sql_keys=sql_keys,
                                          alchemy_url=docassemble.webapp.user_database.alchemy_url,
                                          default_table_class=DEFAULT_TABLE_CLASS,
-                                         default_thead_class=DEFAULT_THEAD_CLASS)
+                                         default_thead_class=DEFAULT_THEAD_CLASS,
+                                         to_text=to_text)
 docassemble.base.functions.set_language(DEFAULT_LANGUAGE, dialect=DEFAULT_DIALECT)
 docassemble.base.functions.set_locale(DEFAULT_LOCALE)
 docassemble.base.functions.update_locale()